Move `ui/comments` state to `comments/ui`
There's no good reason for placing comment-related UI code under the
`ui` sub-tree anymore. With modularisation, it now makes more sense for
it to live with the rest of comments state.
