+++ client/state/comments/ui/test/actions.js
@@ -7,7 +7,7 @@ import { expect } from 'chai';
  * Internal dependencies
  */
 import { COMMENTS_QUERY_UPDATE } from 'state/action-types';
-import { updateCommentsQuery } from 'state/ui/comments/actions';
+import { updateCommentsQuery } from 'state/comments/ui/actions';
 
 describe( 'actions', () => {
 	describe( '#updateCommentsQuery()', () => {
