+++ client/state/selectors/has-pending-comment-requests.js
@@ -4,13 +4,18 @@
 import { get, some } from 'lodash';
 
 /**
+ * Internal dependencies
+ */
+import 'state/comments/init';
+
+/**
  * Returns true if we have any pending comment actions that we are tracking.
  *
  * @param {object} state - global application state
  * @returns {boolean} - true if we have pending actions
  */
 export default ( state ) => {
-	const pendingActions = get( state, 'ui.comments.pendingActions' );
+	const pendingActions = get( state, [ 'comments', 'ui', 'pendingActions' ] );
 	return some( pendingActions, ( requestKey ) => {
 		return get( state, [ 'dataRequests', requestKey, 'status' ] ) === 'pending';
 	} );
