+++ client/state/selectors/test/has-pending-comment-requests.js
@@ -34,7 +34,7 @@ const actionKey2 = getRequestKey( {
 describe( 'hasPendingCommentRequests()', () => {
 	test( 'should return true if we have pending actions', () => {
 		const state = deepFreeze( {
-			ui: { comments: { pendingActions: [ actionKey, actionKey2 ] } },
+			comments: { ui: { pendingActions: [ actionKey, actionKey2 ] } },
 			dataRequests: {
 				[ actionKey ]: { status: 'success' },
 				[ actionKey2 ]: { status: 'pending' },
@@ -44,7 +44,7 @@ describe( 'hasPendingCommentRequests()', () => {
 	} );
 	test( 'should return false if do not have pending actions', () => {
 		const state = deepFreeze( {
-			ui: { comments: { pendingActions: [ actionKey, actionKey2 ] } },
+			comments: { ui: { pendingActions: [ actionKey, actionKey2 ] } },
 			dataRequests: {
 				[ actionKey ]: { status: 'success' },
 				[ actionKey2 ]: { status: 'success' },
@@ -54,7 +54,7 @@ describe( 'hasPendingCommentRequests()', () => {
 	} );
 	test( 'only checks against actions we track in ui state', () => {
 		const state = deepFreeze( {
-			ui: { comments: { pendingActions: [ actionKey ] } },
+			comments: { ui: { pendingActions: [ actionKey ] } },
 			dataRequests: {
 				[ actionKey ]: { status: 'success' },
 				[ actionKey2 ]: { status: 'pending' },
